Combine index.js and main.js into one to improve UX while loading rustpython
